+++ conf/formsStack.conf
@@ -4,7 +4,7 @@ modules = [
   "formsKnot=io.knotx.forms.core.FormsKnot"
 
   ####### Forms Adpater #######
-  "formsAdapter=io.knotx.forms.adapter.example.FormsExampleAdapter"
+  # define here formsAdapter, e.g.: "formsAdapter=your.adapter.example.FormsExampleAdapter"
 ]
 
 ########### Globals ###########
@@ -12,7 +12,7 @@ global {
   address {
     forms {
       knot = knotx.knot.forms
-      example.adapter = knotx.forms.example.adapter.http
+      test.adapter = knotx.forms.it.test.adapter
     }
   }
 }
@@ -25,8 +25,9 @@ config.formsKnot {
 }
 
 ########### Adapter configurations ###########
-config.formsAdapter {
-  options.config {
-    address = ${global.address.forms.example.adapter}
-  }
-}
+# example adapter config
+# config.formsAdapter {
+#   options.config {
+#     address = ${global.address.forms.example.adapter}
+#   }
+# }
