+++ conf/includes/forms.conf
@@ -30,21 +30,5 @@ adapters = [
     #
     # allowedResponseHeaders = []
   # }
-
-  ##### Example forms adapter #####
-  {
-    name = action-self
-    address = ${global.address.forms.example.adapter}
-    params = {
-      example = example-value
-    }
-    allowedRequestHeaders = [
-      "Content-Type"
-      "X-*"
-    ]
-    allowedResponseHeaders = [
-      "X-*"
-    ]
-  }
 ]
