--- it-test/src/main/java/io/knotx/forms/adapter/example/FormsExampleAdapter.java
@@ -1,58 +1,0 @@
-/*
- * Copyright (C) 2018 Knot.x Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.knotx.forms.adapter.example;
-
-
-import io.knotx.forms.api.FormsAdapterProxy;
-import io.vertx.core.Context;
-import io.vertx.core.Vertx;
-import io.vertx.core.eventbus.MessageConsumer;
-import io.vertx.core.json.JsonObject;
-import io.vertx.core.logging.Logger;
-import io.vertx.core.logging.LoggerFactory;
-import io.vertx.reactivex.core.AbstractVerticle;
-import io.vertx.serviceproxy.ServiceBinder;
-
-public class FormsExampleAdapter extends AbstractVerticle {
-
-  private static final Logger LOGGER = LoggerFactory.getLogger(FormsExampleAdapter.class);
-
-  private MessageConsumer<JsonObject> consumer;
-
-  private ServiceBinder serviceBinder;
-
-  @Override
-  public void init(Vertx vertx, Context context) {
-    super.init(vertx, context);
-  }
-
-  @Override
-  public void start() throws Exception {
-    LOGGER.info("Starting <{}>", this.getClass().getSimpleName());
-
-    serviceBinder = new ServiceBinder(getVertx());
-    consumer = serviceBinder
-        .setAddress(config().getString("address"))
-        .register(FormsAdapterProxy.class, new FormsExampleAdapterProxy(
-            config().getString("testStrategy")));
-  }
-
-  @Override
-  public void stop() throws Exception {
-    serviceBinder.unregister(consumer);
-  }
-
-}
