+++ it-test/src/test/java/io/knotx/forms/test/integration/util/KnotContextFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.forms.core.util;
+package io.knotx.forms.test.integration.util;
 
 import io.knotx.dataobjects.ClientRequest;
 import io.knotx.dataobjects.ClientResponse;
