--- it-test/src/test/resources/adapterPost.conf
@@ -1,6 +1,0 @@
-### Override Example Adapter Configuration ###
-config.formsAdapter {
-  options.config {
-    testStrategy = POST_REQUEST
-  }
-}
