+++ it-test/src/test/resources/fragments/form_self_out.html
